Refactor: Move profile hooks to a dedicated directory and update imports
